+++ README.md
@@ -28,6 +28,25 @@ local in this case - could also run one in a docker container
 
 `$ redis-server --protected-mode no`
 
+# stack operations
+
+## deploy stack
+
+### bootstrap cdk resources (staging bucket, etc) once
+`cdk bootstrap`
+
+### list defined stacks in the app - can have stacks defined across multiple accounts, regions, etc 
+`cdk ls`
+
+### deploy stack
+`cdk deploy` 
+
+## destroy stack
+
+`cdk destroy`
+
+# Test deployed stack
+
 ## invoke deployed lambda
 ```shell script
 $ aws lambda invoke \
@@ -41,17 +60,19 @@ The log is base64-encoded in case you need to view it.
 
 Output from lambda is in `lambda.out` 
 
-## bootstrap cdk resources (staging bucket, etc)
-`cdk bootstrap`
+## external api server
 
-## list defined stacks in the app - can have stacks defined across multiple accounts, regions, etc 
-`cdk ls`
+```
+$ curl https://e3ryyap5mf.execute-api.us-east-1.amazonaws.com/prod/survey/1/interview/1/attachment
 
-## Deploy stack
-`cdk deploy` 
+{"data": {"data": "something"}, "quota": 5, "status_code": 200, "message": ["Hello, CDK!  You have hit /survey/1/interview/1/attachment\n"]}
+```
 
-## Destroy stack
-`cdk destroy`
+```bash
+$ curl https://e3ryyap5mf.execute-api.us-east-1.amazonaws.com/prod//list_keys
+
+{"keys": [{"name": "survey:interview:attachment:none", "value": "4", "ttl": 57}, {"name": "survey:interview:none", "value": "9", "ttl": 57}, {"name": "survey:none", "value": "19", "ttl": 57}]}
+```
 
 ## Information generated by cdk init 
 # Welcome to your CDK Python project!
