--- api_poc/api_poc_stack.py
@@ -1,17 +1,0 @@
-from aws_cdk import (
-    core
-)
-
-from external_api_construct import ExternalApi
-
-
-class ApiPocStack(core.Stack):
-    """
-    external api server stack
-    """
-    def __init__(self, scope: core.Construct, id: str, **kwargs) -> None:
-        """ create our stack """
-        super().__init__(scope, id, **kwargs)
-
-        # external api construct
-        ExternalApi(self, 'ExternalApi')
